RavenDB-24796 Fixes after rebase and cleanup
