+++ src/Raven.Studio/typescript/components/pages/database/aiHub/aiAgents/edit/hooks/useEditAiAgent.ts
@@ -1,17 +1,15 @@
 import { databaseSelectors } from "components/common/shell/databaseSliceSelectors";
-import { licenseSelectors } from "components/common/shell/licenseSlice";
 import { useAppUrls } from "components/hooks/useAppUrls";
 import { useDirtyFlag } from "components/hooks/useDirtyFlag";
 import useResizableWidth from "components/hooks/useResizableWidth";
 import { useServices } from "components/hooks/useServices";
 import { connectionStringsActions } from "components/pages/database/settings/connectionStrings/store/connectionStringsSlice";
-import { defaultItemsToProcess } from "components/pages/database/settings/documentExpiration/DocumentExpiration";
 import { useAppDispatch, useAppSelector } from "components/store";
 import { tryHandleSubmit } from "components/utils/common";
 import { useEffect } from "react";
 import { useAsyncCallback } from "react-async-hook";
 import { useForm, useWatch, SubmitHandler } from "react-hook-form";
-import { editAiAgentSelectors, editAiAgentActions } from "../store/editAiAgentSlice";
+import { editAiAgentActions } from "../store/editAiAgentSlice";
 import { editAiAgentUtils } from "../utils/editAiAgentUtils";
 import {
     EditAiAgentFormData,
@@ -20,7 +18,6 @@ import {
     testAiAgentYupResolver,
 } from "../utils/editAiAgentValidation";
 import router from "plugins/router";
-import { TimeInSeconds } from "common/constants/timeInSeconds";
 
 interface QueryParams {
     id: string;
@@ -29,12 +26,10 @@ interface QueryParams {
 
 export default function useEditAiAgent(queryParams: QueryParams) {
     const dispatch = useAppDispatch();
-    const { aiAgentService, databasesService } = useServices();
+    const { aiAgentService } = useServices();
     const { appUrl } = useAppUrls();
 
     const databaseName = useAppSelector(databaseSelectors.activeDatabaseName);
-    const isDocumentExpirationEnabled = useAppSelector(editAiAgentSelectors.isDocumentExpirationEnabled);
-    const isCommunityLicense = useAppSelector(licenseSelectors.licenseType) === "Community";
 
     const isEditAiAgent = !!queryParams?.id && !queryParams.isClone;
 
@@ -48,15 +43,11 @@ export default function useEditAiAgent(queryParams: QueryParams) {
     }, []);
 
     const asyncGetEditDefaultValues = useAsyncCallback(async () => {
-        const isDocumentExpirationEnabled = await dispatch(
-            editAiAgentActions.getIsDocumentExpirationEnabled(databaseName)
-        ).unwrap();
-
         if (queryParams?.id) {
             const agents = await aiAgentService.getAiAgents(databaseName, queryParams.id);
-            return editAiAgentUtils.mapFromDto(agents.AiAgents[0], queryParams.isClone, isDocumentExpirationEnabled);
+            return editAiAgentUtils.mapFromDto(agents.AiAgents[0], queryParams.isClone);
         } else {
-            return editAiAgentUtils.mapFromDto(null, false, isDocumentExpirationEnabled);
+            return editAiAgentUtils.mapFromDto(null, false);
         }
     });
 
@@ -101,22 +92,7 @@ export default function useEditAiAgent(queryParams: QueryParams) {
 
     const saveAgent: SubmitHandler<EditAiAgentFormData> = async (formData) => {
         return tryHandleSubmit(async () => {
-            if (
-                isDocumentExpirationEnabled.status === "success" &&
-                !isDocumentExpirationEnabled.data &&
-                formData.isEnableDocumentExpiration
-            ) {
-                await databasesService.saveExpirationConfiguration(databaseName, {
-                    Disabled: false,
-                    DeleteFrequencyInSec: isCommunityLicense ? minimumCommunityDeleteFrequencyInSec : null,
-                    MaxItemsToProcess: defaultItemsToProcess,
-                });
-            }
-
-            await aiAgentService.saveAiAgent(
-                databaseName,
-                editAiAgentUtils.mapToDto(formData, isDocumentExpirationEnabled.data)
-            );
+            await aiAgentService.saveAiAgent(databaseName, editAiAgentUtils.mapToDto(formData));
 
             editForm.reset(formData);
             setIsDirty(false);
@@ -135,5 +111,3 @@ export default function useEditAiAgent(queryParams: QueryParams) {
         isEditAiAgent,
     };
 }
-
-const minimumCommunityDeleteFrequencyInSec = TimeInSeconds.Day * 36;
